Add modal clocks option
One idea how to work with clocks is by using a modal approach where
certain parts of the central clock schedules the interpolation from a
start to an end pitch. This commit adds basic functionality for this to
work. Most of it is quite dirty, but should be useable in a prototype.