modal: Differentiate between ModalEvent 0 and 1
We differentiate between two different modal movements: ModalEvent0 and ModalEvent1.

    A: ModalEvent0 describes the interpolation between two pitches.
    B: ModalEvent1 describes the static state of one pitch.

So both events are simply different representations of the same reality.
They comprehend each other, since some musical ideas are easier to describe with
respect to ModalEvent0 and some are easier to describe with respect to ModalEvent1.

Before this patch we only had the representation ModalEvent0.
Now we have both.