converters/abjad: Fix half-rendered score
It turns out that the real problem why Lilypond sometimes can't render
the full score seems to be due to rounding errors. On one hand we need
to limit the denominator, because Lilypond doesn't accept infinitely
large denominators. On the other hand we really need to be careful to
pick a high enough number to be sure no rounding errors occur which
result in Lilypond not rendering the full score.